Extract a part of a string, defined as regular expression. Copied directly from DescTools::StrExtract()

Extract a part of a string, defined as regular expression. Copied directly from DescTools::StrExtract()

StrExtract(x, pattern, ...)

Arguments

x
a character vector where matches are sought, or an object which can be coerced by as.character to a character vector.

pattern
character string containing a regular expression (or character string for fixed = TRUE) to be matched in the given character vector. Coerced by as.character to a character string if possible. If a character vector of length 2 or more is supplied, the first element is used with a warning. Missing values are not allowed.

...
the dots are passed to the the internally used function regexpr(), which allows to use e.g. Perl-like regular expressions.